Use another set of patterns.
Former-commit-id: df7982e2050fec8f53cc83be0e69724ff478a896

+++ cpp/type.h
@@ -45,7 +45,7 @@ typedef const line_type& cline_type;
 typedef short pos_type;
 typedef const pos_type& cpos_type;
 
-typedef float val_type;
+typedef double val_type;
 typedef const val_type& cval_type;
 
 enum chessman{blank, white, black, null};
